organize client request tests
